improvement with a composable
